Make widget tests run in web too
