+++ src/test/datascience/notebook/helper.node.ts
@@ -8,7 +8,7 @@ import * as path from '../../../platform/vscode-path/path';
 import * as tmp from 'tmp';
 import { IDisposable } from '../../../platform/common/types';
 import { noop, swallowExceptions } from '../../../platform/common/utils/misc';
-import { EXTENSION_TEST_DIR_FOR_FILES, IS_REMOTE_NATIVE_TEST } from '../../constants.node';
+import { IS_REMOTE_NATIVE_TEST } from '../../constants.node';
 import { commands, NotebookDocument, Uri } from 'vscode';
 import { traceInfo } from '../../../platform/logging';
 import { JupyterServer } from '../jupyterServer.node';
@@ -28,30 +28,6 @@ export async function createTemporaryFile(options: {
         await fs.writeFile(tempFile, options.contents);
     }
     return { file: tempFile, dispose: () => swallowExceptions(() => fs.unlinkSync(tempFile)) };
-}
-
-export async function createTemporaryNotebookFromFile(
-    templateFile: string,
-    disposables: IDisposable[],
-    kernelName: string = 'Python 3'
-): Promise<Uri> {
-    const extension = path.extname(templateFile);
-    fs.ensureDirSync(EXTENSION_TEST_DIR_FOR_FILES);
-    const tempFile = tmp.tmpNameSync({
-        postfix: extension,
-        dir: EXTENSION_TEST_DIR_FOR_FILES,
-        prefix: path.basename(templateFile, '.ipynb')
-    });
-    if (await fs.pathExists(templateFile)) {
-        const contents = JSON.parse(await fs.readFile(templateFile, { encoding: 'utf-8' }));
-        if (contents.kernel) {
-            contents.kernel.display_name = kernelName;
-        }
-        await fs.writeFile(tempFile, JSON.stringify(contents, undefined, 4));
-    }
-
-    disposables.push({ dispose: () => swallowExceptions(() => fs.unlinkSync(tempFile)) });
-    return Uri.file(tempFile);
 }
 
 export async function startJupyterServer(notebook?: NotebookDocument, useCert: boolean = false): Promise<any> {
